Feat. add useMutatePostDocument
