+++ src/components/Modals/AddNewDocument/AddDocModal.jsx
@@ -1,17 +1,7 @@
-import { useState } from "react";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { useAtom, useSetAtom, useAtomValue } from "jotai";
+import { useSetAtom } from "jotai";
 
-import fetchData from "../../../utils/axios";
-
-import {
-  currentDBIdAtom,
-  currentDocIndexAtom,
-  documentsIdsAtom,
-  userAtom,
-  showAddDocumentModalAtom,
-  fieldsAtom,
-} from "../../../atoms/atoms";
+import { showAddDocumentModalAtom } from "../../../atoms/atoms";
+import useMutatePostDocument from "../../../apis/useMutatePostDocument";
 
 import Button from "../../shared/Button";
 import Modal from "../../shared/Modal";
@@ -20,56 +10,10 @@ import AddDocInputList from "./AddDocInputList";
 import ContentWrapper from "../SharedItems/ContentWrapper";
 import Content from "../SharedItems/Content";
 import InputsArea from "../SharedItems/InputsArea";
-import Loading from "../../shared/Loading";
 
 function AddDocumentModal() {
-  const queryClient = useQueryClient();
-
-  const [documentsIds, setDocumentsIds] = useAtom(documentsIdsAtom);
-
-  const { userId } = useAtomValue(userAtom);
-  const fields = useAtomValue(fieldsAtom);
-  const currentDBId = useAtomValue(currentDBIdAtom);
-
-  const setCurrentDocIndex = useSetAtom(currentDocIndexAtom);
+  const fetchNewDocument = useMutatePostDocument();
   const setShowAddDocumentModal = useSetAtom(showAddDocumentModalAtom);
-
-  function addNewDocumentId(newId) {
-    const newFields = [...documentsIds];
-
-    newFields.push(newId);
-    setDocumentsIds(newFields);
-  }
-
-  async function handleClickSave() {
-    const response = await fetchData(
-      "POST",
-      `/users/${userId}/databases/${currentDBId}/documents`,
-      fields,
-    );
-
-    return response.data.newDocument._id;
-  }
-
-  const { mutate: fetchDocumentSave, isLoading } = useMutation(
-    handleClickSave,
-    {
-      onSuccess: result => {
-        addNewDocumentId(result);
-        setCurrentDocIndex(documentsIds.length);
-
-        queryClient.refetchQueries(["dbDocumentList", currentDBId]);
-        setShowAddDocumentModal(false);
-      },
-      onFailure: () => {
-        console.log("sending user to errorpage");
-      },
-    },
-  );
-
-  if (isLoading) {
-    return <Loading />;
-  }
 
   return (
     <Modal onClick={() => setShowAddDocumentModal(false)}>
@@ -82,7 +26,7 @@ function AddDocumentModal() {
         </Content>
         <Button
           className="w-20 h-8 mt-5 rounded-md bg-black-bg text-white hover:bg-dark-grey"
-          onClick={fetchDocumentSave}
+          onClick={() => fetchNewDocument()}
         >
           Submit
         </Button>
